% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{dgp_f}
\alias{dgp_f}
\title{Deterministic target function f(x) for the project DGP}
\usage{
dgp_f(x, type = c("1d", "2d"))
}
\arguments{
\item{x}{A numeric vector (for 1d) or a numeric matrix with 2 columns (for 2d).}

\item{type}{Character. One of \code{"1d"} or \code{"2d"}.}
}
\value{
A numeric vector of function values.
}
\description{
Two target functions are supported:
\itemize{
\item \code{"1d"}: \eqn{f(x) = \sin(2\pi x)}
\item \code{"2d"}: \eqn{f(x_1,x_2) = \sin(\sqrt{x_1^2 + x_2^2})}
}
}
\examples{
f1 <- dgp_f(seq(0, 1, length.out = 5), type = "1d")
x2 <- cbind(c(0, 1), c(0, 1))
f2 <- dgp_f(x2, type = "2d")
f1; f2

}
