% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{knn_predict}
\alias{knn_predict}
\title{k-NN regression prediction at test points}
\usage{
knn_predict(x_train, y_train, x_test, k)
}
\arguments{
\item{x_train}{Numeric matrix of training covariates (n x d).}

\item{y_train}{Numeric vector of training responses (length n).}

\item{x_test}{Numeric matrix of test covariates (m x d) or numeric vector (length m) for 1d.}

\item{k}{Integer. Number of neighbors.}
}
\value{
Numeric vector of predictions of length m.
}
\description{
Computes k-nearest neighbors regression predictions:
\deqn{\hat f(x_0) = \frac{1}{k}\sum_{i \in \mathcal{N}_k(x_0)} y_i}
where neighbors are defined by Euclidean distance.
}
\examples{
x <- matrix(c(0, 0.5, 1), ncol = 1)
y <- c(0, 1, 2)
pred <- knn_predict(x, y, x_test = c(0.1, 0.9), k = 1)
pred

}
