% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{make_grid}
\alias{make_grid}
\title{Create an evaluation grid on \link{0, 1}^d}
\usage{
make_grid(grid_size = 101, d = 1)
}
\arguments{
\item{grid_size}{Integer. For d=1, number of points. For d=2, points per axis (total grid_size^2).}

\item{d}{Integer. Dimension (1 or 2).}
}
\value{
A list with:
\itemize{
\item \code{x_grid}: numeric matrix of grid points.
\item \code{axis}: numeric vector used for each axis (useful for plotting in 2d).
}
}
\description{
Create an evaluation grid on \link{0, 1}^d
}
\examples{
g1 <- make_grid(grid_size = 5, d = 1)
g2 <- make_grid(grid_size = 5, d = 2)
nrow(g1$x_grid); nrow(g2$x_grid)

}
